Blog post 7 incorporated: Remap the Kernel
This is a crazy large update, it is pretty much verbatim
taken from Phillip's blog so it is better to just look
there for the down-low on what happened. I have one TODO
from this post and that is to identity map the multiboot2
kernel modules, but that requires mapping being availible
to lib.rs which is so far not true.