Finished the blog
So we don't follow the blog 100% of course, but at this
point our repository is quite similar to Phillip's. I'm
not going to go through and explain all of my changes,
as one can just check out the blog to figure out what
I have done, but I will try to explain the differences.
The main diversions are in the interrupts submodule
because that is what I have been working on by myself
without the blog. By suggestion of people on #osdev
and #rust-osdev I am keeping isr stubs in
`src/arch/x86_64/interrupts.asm`, and have improved the
macros in that file, contrary to the blog where he uses
macros in the interrupts file. We also have keyboard
support.
Moving on I would like to make the kernel higher half,
bulk up the memory module so that mapping pages from
other modules is easy, and eventually lead up to a
kernel space -> userspace ring transfer.

From now on, It's all us.