Link kernel to the higher half
This commit was pretty much a rewrite of `boot.asm`, this included
the consolodation of `long_mode_init.asm` which is now just part of
that file. All of the changes have to do with how the kernel is
linked. Now one section, `.init` is linked at 1M. This sets up paging
and maps the rest of the kernel to the higher half. The rest of the
kernel code is linked to 1M + 0xFFFFFFF80000000, although it is still
allocated at 1M. All memory before 1M is identity mapped so that
accessing memory mapped hardware will be easier. This commit also
added some miscellaneous documentation improvements for optimal
readability of `boot.asm`

The commit also made much of the written rust code useless, and it
will need to be rewritten before it can be used again. This can be
seen in my changes to `lib.rs` which comment out much of the code.
Specifically the multiboot2 information structure (and everything
that it points to) is useless which renders much of ::memory entirely
useless. This will need to be changed before this commit is merged
into Master.