Make multiboot2 information structure usable
It seems like the multiboot2 crate does not really
dereference very much of the information structure
by itself, so it is usable at the moment. We will
have to take great caution when reading the elf-sections
tag because most of the information will be somewhat off,
but the multiboot2 modules seem to work very well right
now. Also many of the control register bit fiddling that
was done in the kernel is mooved to `boot.asm` for
simplicity and efficiency. Memory does not yet work and
is therefore still commented out.