Make Rust code higher half viable
This commit mostly deals with small changes in the ::memory::paging
module, although there were certain changes outside of that. We are
no longer using the git version of multiboot (and all of the bad
things that came with that) and are using version "0.3.0" which has
a relatively stable interface that works well and does what we want.
I also fixed some minor (non catostrophic) bugs that were commited
during the late-night existential drama that was the conversion to
higher half. Along with that I have tried to remove warning causes,
such as the never really needed use of #![feature(core_intrinsics)]

The features availible at this commit should reflect those before
the kernel migrated to higher half, as the entire purpose of it was
a bug fix, not a feature addition.