Fix TinyAllocator frame leakage
This commit adds a hack that requires TemporaryPage's `drop()`
method to be run at the end of it's lifetime. I was not able to
make `drop()` consume the page, however the TinyAllocator's behavior
will not be undefined if it is used after that method is called. It
will simply not have any frames to allocate.
This should probably change in the future, but the frame leakage
was stopped as is desired.