Improves documentation, comments, and types
This commit adds documentation to practically every function in the
repository. This can be rendered with `rustdoc` but it is currently
not working optimally because only the public interface is shown.
There is still a way to generate all documentation, and I am working
on a way to automate it.