Fix blocking and recursion in StackFrameAllocator
This commit fixes the blocking in StackFrameAllocator. There is also
blocking in `remap_the_kernel()`, and that was fixed using a temporary
change. While fixing things I added more verbose comments to some
functions and errors in the operating system. `test_paging` has been
moved from being a dead function to being commented out, it will be
returned to the codebase when the memory module is fixed.

The biggest changes in this commit are the changes to `TinyAllocator`.
Now almost all of it's functions are generic using closures which makes
it easier to use in other parts of the codebase.

Finally I have created some changes in the `vga_buffer` module. These
are mostly whitespace fixes, however I have improved functionality in
some parts.