Fix compilation and generally improve things
This project hasn't had a commit in awhile, so this is more of a messy
sweep-up job than anything else.
Months of negligence on a "nightly" versioned project leads to several
compilation failures, which have now been fixed. These include the
merging of `collections` and `alloc` crates. The change of the `Unique`
struct. And a new target option has been added, which made cross cross
compilation fail.
There have been also several general improvements. These include the new
use of the `x86_64` crate, instead of x86, which simplifies code (but
also requires a quite larger stack). This enables interrupt handling to
be greatly simplified when paired with the new "x86-interrupt" abi. Also
there is a small change to printing that allows for waiting for the
writer to flush. This is very useful for color switching.
Finally there have been a few minor changes, including changing the name
of the ugly `(ARCH)-unknown-none-gnu` to `(ARCH)-ESALP`. And a small
.gdbinit file has been added which should simplify debugging with qemu.

Many of these changes are better shown my themselves in Blog-OS. Please
see phil-opp/blog_os#289 phil-opp/blog_os#325 and phil-opp/blog_os#333
for more details.