Use a bitmap for allocating physical frames
This commit swaps out the `StackFrameAllocator` for a struct that
allocates physical frames in a bitmap. This has several advantages (and
fewer disadvantages). The stack allocator was very difficult to deal
with because it needed sole control of the active table, this struct can
be used in conjunction with it, only requiring allocation for its
constructor. Where `n` is the number of physical frames, the bitmap
requires `n/8` space and the stack requires `n*8`, that means we get a
64x difference by changing allocators! The bitmap cannot be deallocated
though, but this is fine because it only uses a one page which I believe
is acceptable. Also this brings ESALP code back closer to blog_os, which
has a rather good memory management that we should look to.
The only major downside to using it is a relatively large cost for
initialization, however I think that the code cleanup completely
justifies this.

Also some refactoring has been done to the interrupts sub module, nothing
major except for the addition of a double fault handler.

This is part of a project for reducing locks and stack buffers in the os
code. The stack had to grow wildly too large in the previous commit, and
I would like to continue to decrease it to 1 or 2 pages at most. Rust
likes to allocate on the stack, so this will probably require re-working
or dropping several crates such as `x66_64`, however I think that it is
very much worth the work. Kernel stacks are very important and should be
used in moderation.