Use separate stack for double faults
This commit adds a `StackAllocator` struct to the memory module. It uses
a fixed area of memory, and allows for any code in the project to create
a stack for its use. The `StackAllocator` can currently not deallocate
or grow stacks, however it should be expanded to do so in the future.
The kernel now uses a Rust managed GDT, and we stop using the one that
was created in boot assembly. We use both these changes to load a new
stack whenever a double fault occurs, this is useful because it will
allow for the double fault message and stack frame to be printed even if
a stack overflow occurs.
No more triple faults should occur after this commit!