Allow non-alphabetic keys to have shift variants
Each variant is hardcoded in a match, a keyboard layout cannot define
its own. This does not allow for keyboard layouts like programmer's
dvorak, which has symbols on different numbers. However, it is unlikely
that layouts like this will be supported before keyboard layout handling
is sent to the userspace.