Make tap.rs use println
By the time tap.rs is used, prining is
available, so we use that. There's no way we're
gonna be able to use it to test printing, but
at this point I think the trade-off is worth the
reduced complexity. Besides, it probably wouldn't
have been able to test printing anyways, since it
was using a lock on the vga buffer.