Create inactive page table
Make a new module for creation of processes,
and in that module, create a new page table
that will become the address space for our
new process

Copy p3 entries to new p4 table

Copy the kernel (in the higher half) into
the new address space. Many memory-module
functions are made public to allow this.