Add copy kernel option to InactivePageTable
Fix the copy_higher_half() method of Table
to map the table before using it

Found bug(?) in recursive mapping of original
table, put in a fixme

Add testing for copy_higher_half()

Derive a few traits to help with debugging of
entries. Also, wrote print_entries() method on
Table

Remove the now unused src/process.rs file

Finally, add the option to copy the higher half
upon creation of an InactivePageTable. This is not
default behavior because of the aforementioned bug.